+++ sandwich-contexts/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for sandwich-contexts
 
+## Unreleased
+
+* Add `postgresNixConfExtraLines` to `PostgresNixOptions`, to be able to add extra lines to `postgresql.conf`.
+
 ## 0.3.0.1
 
 * Automatically find sufficiently short Unix socket paths when the system temp ones are too long